requirements.txt
aiogram>=3.6.0
aiocryptopay>=0.4.0
aiohttp>=3.9.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
